Move all remaining resource types from core-types to graphene-resources
