+++ node-graph/nodes/raster/Cargo.toml
@@ -17,6 +17,7 @@ std = [
 	"serde",
 	"dep:core-types",
 	"dep:dyn-any",
+	"dep:graphene-resource",
 	"dep:graphene-hash",
 	"dep:raster-types",
 	"dep:vector-types",
@@ -43,6 +44,7 @@ node-macro = { workspace = true }
 # Local std dependencies
 dyn-any = { workspace = true, optional = true }
 core-types = { workspace = true, optional = true }
+graphene-resource = { workspace = true, optional = true }
 graphene-hash = { workspace = true, optional = true }
 raster-types = { workspace = true, optional = true }
 vector-types = { workspace = true, optional = true }
